change get commit counts to line changes count

+++ src/main/java/com/hzhang/whotouchedwhat/model/Author.java
@@ -6,10 +6,10 @@ public class Author {
     @JsonProperty("author_name")
     String name;
     @JsonProperty("commit_count")
-    int commitCount;
+    long commitCount;
     @JsonProperty("y_axis")
     String yAxis;
-    public Author(String name, int count) {
+    public Author(String name, long count) {
         this.name = name;
         commitCount = count;
         yAxis = "Authors";
